refactor(seed): establish domain ownership seam
