+++ packages/compiler/src/seed/index.ts
@@ -1,2 +1,15 @@
+export {
+	createCommittedSeed,
+	orderCommittedSeeds,
+	verifyCommittedSeed,
+} from "./committed-seed";
+export type {
+	CommittedSeedV1,
+	SeedFieldValueV1,
+	SeedStepV1,
+	SeedValueV1,
+} from "./committed-seed";
 export { loadCommittedSeed } from "./artifact-files";
+export { applyCommittedSeeds } from "./postgres/apply";
+export type { ApplySeedsResult } from "./postgres/apply";
 export { validateCommittedSeedSchema } from "./schema-validation";
